Omitting any extra newlines in the pilet's preamble to maintain the validity of original Parcel sourcemaps.
The postprocessing of the bundling action and also the loading (via "new Function()") of pilets itself was producing extra newlines in the beginning of the generated code. This was causing a mismatch against sourcemaps that were produced earlier by Parcel. Therefore extra generated code is added to the bundle by using semicolons (instead of newlines) as a separator now.